+++ src/app/components/PersonalAvatar/types.ts
@@ -1,4 +1,7 @@
-import { AvatarSize } from "../Avatar/types"
+import { AvatarSize } from '../Avatar/types';
+
+/** Props del componente PersonalAvatar */
 export interface PersonalAvatarProps {
-    size?:AvatarSize
+  /** Tamaño del avatar mostrado */
+  size?: AvatarSize;
 }